dont depend on rocksdb crate in adapter/sql
