+++ src/adapter/Cargo.toml
@@ -38,7 +38,7 @@ mz-pgrepr = { path = "../pgrepr" }
 mz-postgres-util = { path = "../postgres-util" }
 mz-proto = { path = "../proto" }
 mz-repr = { path = "../repr", features = ["tracing_"] }
-mz-rocksdb = { path = "../rocksdb" }
+mz-rocksdb-types = { path = "../rocksdb-types" }
 mz-secrets = { path = "../secrets" }
 mz-segment = { path = "../segment" }
 mz-sql = { path = "../sql" }
