+++ src/rocksdb-types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "mz-rocksdb"
-description = "A wrapper for RocksDB."
+name = "mz-rocksdb-types"
+description = "Shared types for the `mz-rocksdb` crate"
 version = "0.0.0"
 edition.workspace = true
 rust-version.workspace = true
@@ -8,26 +8,13 @@ publish = false
 
 [dependencies]
 anyhow = "1.0.66"
-bincode = { version = "1.3.3" }
-itertools = { version = "0.10.5" }
-once_cell = "1.16.0"
 mz-ore = { path = "../ore", features = ["async", "metrics", "test"] }
 mz-proto = { path = "../proto" }
-mz-rocksdb-types = { path = "../rocksdb-types" }
 num_cpus = "1.14.0"
 prost = { version = "0.11.3", features = ["no-recursion-limit"] }
 proptest = { version = "1.0.0", default-features = false, features = ["std"]}
 proptest-derive = { version = "0.3.0", features = ["boxed_union"]}
-tokio = { version = "1.24.2", features = ["macros", "sync", "rt"] }
 serde = { version = "1.0.152", features = ["derive"] }
-serde_json = { version = "1.0.89" }
-thiserror = "1.0.37"
-tracing = "0.1.37"
-# These features use compression code that are licensed with:
-# https://github.com/google/snappy/blob/main/COPYING
-# https://github.com/lz4/lz4/blob/dev/LICENSE
-# https://github.com/facebook/zstd
-rocksdb = { version = "0.21.0", default-features = false, features = ["snappy", "zstd", "lz4"] }
 uncased = "0.9.7"
 workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
 
@@ -35,10 +22,6 @@ workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
 prost-build = "0.11.2"
 protobuf-src = "1.1.0"
 tonic-build = "0.8.2"
-
-[dev-dependencies]
-tempfile = "3.2.0"
-prometheus = { version = "0.13.3", default-features = false }
 
 [package.metadata.cargo-udeps.ignore]
 normal = ["workspace-hack"]
