+++ src/rocksdb-types/build.rs
@@ -89,6 +89,6 @@ fn main() {
         // broad, but it's better.
         .emit_rerun_if_changed(false)
         .extern_path(".mz_proto", "::mz_proto")
-        .compile_with_config(config, &["rocksdb/src/config.proto"], &[".."])
+        .compile_with_config(config, &["rocksdb-types/src/config.proto"], &[".."])
         .unwrap_or_else(|e| panic!("{e}"))
 }
